+++ nengo/builder/connection.py
@@ -2,7 +2,7 @@ import collections
 
 import numpy as np
 
-from nengo.builder import Builder, Signal
+from nengo.builder import Builder
 from nengo.builder.ensemble import gen_eval_points, get_activities
 from nengo.builder.node import SimPyFunc
 from nengo.builder.operator import Copy, ElementwiseInc
@@ -62,19 +62,21 @@ def get_eval_points(model, conn, rng):
     if conn.eval_points is None:
         view = model.params[conn.pre_obj].eval_points.view()
         view.setflags(write=False)
+        assert view.dtype == model.dtype
         return view
     else:
         return gen_eval_points(
-            conn.pre_obj, conn.eval_points, rng, conn.scale_eval_points)
+            conn.pre_obj, conn.eval_points, rng, conn.scale_eval_points,
+            dtype=model.dtype)
 
 
-def get_targets(conn, eval_points):
+def get_targets(conn, eval_points, dtype=np.float64):
     if conn.function is None:
-        targets = eval_points[:, conn.pre_slice]
+        targets = eval_points[:, conn.pre_slice].astype(dtype)
     elif isinstance(conn.function, np.ndarray):
         targets = conn.function
     else:
-        targets = np.zeros((len(eval_points), conn.size_mid))
+        targets = np.zeros((len(eval_points), conn.size_mid), dtype=dtype)
         for i, ep in enumerate(eval_points[:, conn.pre_slice]):
             out = conn.function(ep)
             if out is None:
@@ -95,7 +97,7 @@ def build_linear_system(model, conn, rng):
             "This is because no evaluation points fall in the firing "
             "ranges of any neurons." % (conn, conn.pre_obj))
 
-    targets = get_targets(conn, eval_points)
+    targets = get_targets(conn, eval_points, dtype=model.dtype)
     return eval_points, activities, targets
 
 
@@ -105,7 +107,7 @@ def build_decoders(model, conn, rng):
     bias = model.params[conn.pre_obj].bias
 
     eval_points = get_eval_points(model, conn, rng)
-    targets = get_targets(conn, eval_points)
+    targets = get_targets(conn, eval_points, dtype=model.dtype)
 
     if conn.solver.weights and not conn.solver.compositional:
         # solver is solving for the whole weight matrix, so apply
@@ -159,7 +161,8 @@ def slice_signal(model, signal, sl):
         return signal[sl]
     else:
         size = np.arange(signal.size)[sl].size
-        sliced_signal = Signal(np.zeros(size), name="%s.sliced" % signal.name)
+        sliced_signal = model.Signal(np.zeros(size, dtype=model.dtype),
+                                     name="%s.sliced" % signal.name)
         model.add_op(Copy(signal, sliced_signal, src_slice=sl))
         return sliced_signal
 
@@ -249,7 +252,8 @@ def build_connection(model, conn):
         elif isinstance(conn.function, np.ndarray):
             raise BuildError("Cannot use function points in direct connection")
         else:
-            in_signal = Signal(np.zeros(conn.size_mid), name='%s.func' % conn)
+            in_signal = model.Signal(np.zeros(conn.size_mid),
+                                     name='%s.func' % conn)
             model.add_op(SimPyFunc(in_signal, conn.function, None, sliced_in))
     elif isinstance(conn.pre_obj, Ensemble):  # Normal decoded connection
         eval_points, decoders, solver_info = model.build(
@@ -296,8 +300,9 @@ def build_connection(model, conn):
     if isinstance(conn.post_obj, Neurons):
         # Apply neuron gains (we don't need to do this if we're connecting to
         # an Ensemble, because the gains are rolled into the encoders)
-        gains = Signal(model.params[conn.post_obj.ensemble].gain[post_slice],
-                       name="%s.gains" % conn)
+        gains = model.Signal(
+            model.params[conn.post_obj.ensemble].gain[post_slice],
+            name="%s.gains" % conn)
         model.add_op(ElementwiseInc(
             gains, weighted, model.sig[conn]['out'][post_slice],
             tag="%s.gains_elementwiseinc" % conn))