+++ nengo/spa/pointer.py
@@ -1,6 +1,7 @@
 import numpy as np
 
 from nengo.exceptions import ValidationError
+from nengo.rc import global_dtype
 from nengo.utils.numpy import is_integer, is_number
 
 
@@ -25,7 +26,7 @@ class SemanticPointer:
                 raise ValidationError(
                     "Must specify either the data or the length for a "
                     "SemanticPointer.", attr='data', obj=self)
-            self.v = np.array(data, dtype=float)
+            self.v = np.array(data, dtype=global_dtype)
             if len(self.v.shape) != 1:
                 raise ValidationError("'data' must be a vector", 'data', self)